(PC-18849)[PRO] refactor: move submit logic next to form
