+++ pro/src/screens/OfferEducational/OfferEducational.tsx
@@ -1,5 +1,6 @@
 import { FormikProvider, useFormik } from 'formik'
 import React, { useEffect } from 'react'
+import { useHistory } from 'react-router-dom'
 
 import { GetEducationalOffererResponseModel } from 'apiClient/v1'
 import OfferEducationalActions from 'components/OfferEducationalActions'
@@ -9,18 +10,29 @@ import {
   IOfferEducationalFormValues,
   EducationalCategories,
   Mode,
+  isCollectiveOffer,
+  isCollectiveOfferTemplate,
+  CollectiveOffer,
+  CollectiveOfferTemplate,
 } from 'core/OfferEducational'
+import patchCollectiveOfferAdapter from 'core/OfferEducational/adapters/patchCollectiveOfferAdapter'
+import postCollectiveOfferAdapter from 'core/OfferEducational/adapters/postCollectiveOfferAdapter'
+import postCollectiveOfferTemplateAdapter from 'core/OfferEducational/adapters/postCollectiveOfferTemplateAdapter'
+import { computeURLCollectiveOfferId } from 'core/OfferEducational/utils/computeURLCollectiveOfferId'
 import { SelectOption } from 'custom_types/form'
+import useNotification from 'hooks/useNotification'
+import { useCollectiveOfferImageUpload } from 'pages/CollectiveOfferCreation/useCollectiveOfferImageUpload'
+import { patchCollectiveOfferTemplateAdapter } from 'pages/CollectiveOfferEdition/adapters/patchCollectiveOfferTemplateAdapter'
 
 import styles from './OfferEducational.module.scss'
 import OfferEducationalForm from './OfferEducationalForm'
-import { IImageUploaderOfferProps } from './OfferEducationalForm/FormImageUploader/FormImageUploader'
 import { validationSchema } from './validationSchema'
 
 export interface IOfferEducationalProps {
+  offer?: CollectiveOffer | CollectiveOfferTemplate
+  setOffer: (offer: CollectiveOffer | CollectiveOfferTemplate) => void
   categories: EducationalCategories
   initialValues: IOfferEducationalFormValues
-  onSubmit(values: IOfferEducationalFormValues): void
   userOfferers: GetEducationalOffererResponseModel[]
   getIsOffererEligible?: CanOffererCreateCollectiveOffer
   mode: Mode
@@ -31,17 +43,15 @@ export interface IOfferEducationalProps {
   isOfferCancellable?: boolean
   domainsOptions: SelectOption[]
   isTemplate: boolean
-  imageOffer: IImageUploaderOfferProps['imageOffer']
-  onImageUpload: IImageUploaderOfferProps['onImageUpload']
-  onImageDelete: IImageUploaderOfferProps['onImageDelete']
   isOfferCreated?: boolean
 }
 
 const OfferEducational = ({
+  offer,
+  setOffer,
   categories,
   userOfferers,
   initialValues,
-  onSubmit,
   getIsOffererEligible,
   mode,
   cancelActiveBookings,
@@ -51,11 +61,72 @@ const OfferEducational = ({
   isOfferActive = false,
   domainsOptions,
   isTemplate,
-  imageOffer,
-  onImageUpload,
-  onImageDelete,
   isOfferCreated = false,
 }: IOfferEducationalProps): JSX.Element => {
+  const notify = useNotification()
+  const history = useHistory()
+  const { imageOffer, onImageDelete, onImageUpload, handleImageOnSubmit } =
+    useCollectiveOfferImageUpload(offer, isTemplate)
+
+  const onSubmit = async (offerValues: IOfferEducationalFormValues) => {
+    let response = null
+    if (isTemplate) {
+      if (offer === undefined) {
+        response = postCollectiveOfferTemplateAdapter({ offer: offerValues })
+      } else {
+        response = await patchCollectiveOfferTemplateAdapter({
+          offer: offerValues,
+          initialValues,
+          offerId: offer.id,
+        })
+      }
+    } else {
+      if (offer === undefined) {
+        response = postCollectiveOfferAdapter({ offer: offerValues })
+      } else {
+        response = await patchCollectiveOfferAdapter({
+          offer: offerValues,
+          initialValues,
+          offerId: offer.id,
+        })
+      }
+    }
+
+    const { payload, isOk, message } = await response
+
+    if (!isOk) {
+      return notify.error(message)
+    }
+
+    const offerId = offer?.id ?? payload.id
+    await handleImageOnSubmit(offerId)
+    if (
+      offer &&
+      (isCollectiveOffer(payload) || isCollectiveOfferTemplate(payload))
+    ) {
+      setOffer({
+        ...payload,
+        imageUrl: imageOffer?.url,
+        imageCredit: imageOffer?.credit,
+      })
+    }
+
+    if (mode === Mode.EDITION && offer !== undefined) {
+      return history.push(
+        `/offre/${computeURLCollectiveOfferId(
+          offer.id,
+          offer.isTemplate
+        )}/collectif/recapitulatif`
+      )
+    }
+
+    history.push(
+      isTemplate
+        ? `/offre/${payload.id}/collectif/vitrine/creation/recapitulatif`
+        : `/offre/${payload.id}/collectif/stocks`
+    )
+  }
+
   const { resetForm, ...formik } = useFormik({
     initialValues,
     onSubmit,
